Using own files for different versions to handle library changes
