+++ csync/csynclib/v0_90_0.py
@@ -1,0 +1,82 @@
+# values for enumeration 'csync_error_codes_e'
+(CSYNC_ERR_NONE,
+CSYNC_ERR_LOG,
+CSYNC_ERR_LOCK,
+CSYNC_ERR_STATEDB_LOAD,
+CSYNC_ERR_STATEDB_WRITE,
+CSYNC_ERR_MODULE,
+CSYNC_ERR_TIMESKEW,
+CSYNC_ERR_FILESYSTEM,
+CSYNC_ERR_TREE,
+CSYNC_ERR_MEM,
+CSYNC_ERR_PARAM,
+CSYNC_ERR_UPDATE,
+CSYNC_ERR_RECONCILE,
+CSYNC_ERR_PROPAGATE,
+CSYNC_ERR_ACCESS_FAILED,
+CSYNC_ERR_REMOTE_CREATE,
+CSYNC_ERR_REMOTE_STAT,
+CSYNC_ERR_LOCAL_CREATE,
+CSYNC_ERR_LOCAL_STAT,
+CSYNC_ERR_PROXY,
+CSYNC_ERR_LOOKUP,
+CSYNC_ERR_AUTH_SERVER,
+CSYNC_ERR_AUTH_PROXY,
+CSYNC_ERR_CONNECT,
+CSYNC_ERR_TIMEOUT,
+CSYNC_ERR_HTTP,
+CSYNC_ERR_PERM,
+CSYNC_ERR_NOT_FOUND,
+CSYNC_ERR_EXISTS,
+CSYNC_ERR_NOSPC,
+CSYNC_ERR_QUOTA,
+CSYNC_ERR_SERVICE_UNAVAILABLE,
+CSYNC_ERR_FILE_TOO_BIG,
+CSYNC_ERR_ABORTED,
+CSYNC_ERR_UNSPEC,) =  xrange(35)
+
+# values for enumeration 'csync_ftw_type_e'
+(CSYNC_FTW_TYPE_FILE,
+CSYNC_FTW_TYPE_SLINK,
+CSYNC_FTW_TYPE_DIR,
+CSYNC_FTW_TYPE_SKIP,) = xrange(4)
+
+# values for enumeration 'csync_instructions_e'
+CSYNC_INSTRUCTION_NONE = 0
+CSYNC_INSTRUCTION_EVAL = 1
+CSYNC_INSTRUCTION_REMOVE = 2
+CSYNC_INSTRUCTION_RENAME = 4
+CSYNC_INSTRUCTION_NEW = 8
+CSYNC_INSTRUCTION_CONFLICT = 16
+CSYNC_INSTRUCTION_IGNORE = 32
+CSYNC_INSTRUCTION_SYNC = 64
+CSYNC_INSTRUCTION_STAT_ERROR = 128
+CSYNC_INSTRUCTION_ERROR = 256
+CSYNC_INSTRUCTION_DELETED = 512
+CSYNC_INSTRUCTION_UPDATED = 1024
+
+# values for enumeration 'csync_notify_type_e'
+(CSYNC_NOTIFY_INVALID,
+CSYNC_NOTIFY_START_SYNC_SEQUENCE,
+CSYNC_NOTIFY_START_DOWNLOAD,
+CSYNC_NOTIFY_START_UPLOAD,
+CSYNC_NOTIFY_PROGRESS,
+CSYNC_NOTIFY_FINISHED_DOWNLOAD,
+CSYNC_NOTIFY_FINISHED_UPLOAD,
+CSYNC_NOTIFY_FINISHED_SYNC_SEQUENCE,
+CSYNC_NOTIFY_START_DELETE,
+CSYNC_NOTIFY_END_DELETE,
+CSYNC_NOTIFY_ERROR,) = xrange(11)
+
+(CSYNC_LOG_PRIORITY_NOLOG,
+CSYNC_LOG_PRIORITY_FATAL,
+CSYNC_LOG_PRIORITY_ALERT,
+CSYNC_LOG_PRIORITY_CRIT,
+CSYNC_LOG_PRIORITY_ERROR,
+CSYNC_LOG_PRIORITY_WARN,
+CSYNC_LOG_PRIORITY_NOTICE,
+CSYNC_LOG_PRIORITY_INFO,
+CSYNC_LOG_PRIORITY_DEBUG,
+CSYNC_LOG_PRIORITY_TRACE,
+CSYNC_LOG_PRIORITY_NOTSET,
+CSYNC_LOG_PRIORITY_UNKNOWN,) = xrange(12)
